requirements.txt
manim
manim-voiceover
manim-voiceover[recorder]
manim-voiceover[transcribe]
matplotlib
pillow
pygments